improve and rearrange node_types plan modifier
